fix hook render
